+++ mkapi/core/object.py
@@ -201,8 +201,21 @@ def get_origin(obj: Any) -> Any:
         return get_origin(obj.fget)
     if not callable(obj):
         return obj
-    if hasattr(obj, "__wrapped__"):
-        return get_origin(obj.__wrapped__)
-    if hasattr(obj, "__pytest_wrapped__"):
-        return get_origin(obj.__pytest_wrapped__.obj)
+    # if hasattr(obj, "__wrapped__"):
+    #     return get_origin(obj.__wrapped__)
+    # if hasattr(obj, "__pytest_wrapped__"):
+    #     return get_origin(obj.__pytest_wrapped__.obj)
+    try:
+        wrapped = obj.__wrapped__
+    except AttributeError:
+        pass
+    else:
+        return get_origin(wrapped)
+    try:
+        wrapped = obj.__pytest_wrapped__
+    except AttributeError:
+        pass
+    else:
+        return get_origin(wrapped.obj)
+
     return obj